fix infinite api calls bug
