+++ app/page.tsx
@@ -1,12 +1,11 @@
-"use client";
-import AddRecipeButton from "@/components/AddRecipeButton/AddRecipeButton";
+import AddRecipe from "@/components/AddRecipe/AddRecipe";
 import Recipes from "@/components/Recipes/Recipes";
 
 export default function Home() {
   return (
     <>
       <Recipes />
-      <AddRecipeButton />
+      <AddRecipe />
     </>
   );
 }
